INTERPOLATION CUBIIIIQUE !

+++ src/obstacles/MoteurPhysique.java
@@ -26,7 +26,7 @@ public class MoteurPhysique implements Service {
 	protected Log log;
 	
 	private int distanceApproximation;
-	private int courbureMax;
+	private double courbureMax;
 	
 	public MoteurPhysique(Log log)
 	{
@@ -144,7 +144,7 @@ public class MoteurPhysique implements Service {
 	@Override
 	public void useConfig(Config config)
 	{
-		courbureMax = config.getInt(ConfigInfo.COURBURE_MAX);
+		courbureMax = config.getDouble(ConfigInfo.COURBURE_MAX);
 		distanceApproximation = config.getInt(ConfigInfo.DISTANCE_MAX_ENTRE_MESURE_ET_OBJET);		
 	}
